+++ autofit/mapper/prior_model/prior_model.py
@@ -10,7 +10,7 @@ from autofit.mapper.model_object import ModelObject
 from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from autofit.mapper.prior_model.deferred import DeferredInstance
 from autofit.mapper.prior_model.prior import TuplePrior, Prior
-from autofit.tools.promise import Promise
+from autofit.mapper.promise.promise import Promise
 from autofit.mapper.prior_model.abstract import check_assertions
 
 logger = logging.getLogger(__name__)
