+++ autofit/mapper/promise/__init__.py
@@ -1,0 +1,2 @@
+from .promise import *
+from .assertion_promise import *
